Portal_Application/requirements.txt
streamlit
playwright
python-dotenv 